// vet_panchito/src/app/models/historiasclinicas.ts
export interface HistoriaClinica {
  id?: number;
  nombreDueno: string;
  correoDueno: string;
  celularDueno: string;
  direccionDueno: string;
  documentoDueno: string;
  numeroDocumentoDueno: string;
  nombrePaciente: string;
  especiePaciente: string;
  razaPaciente: string;
  edadPaciente: string;
  pesoPaciente: string;
  sexoPaciente: string;
  enfermedadesPaciente: string;
  resenaPaciente: string;
}